add-tag

+++ app/Http/Controllers/Admin/PostController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\Admin;
 
+use App\Tag;
 use App\Post;
 use App\Category;
 use Illuminate\Http\Request;
@@ -11,6 +12,21 @@ use Illuminate\Support\Facades\Storage;
 
 class PostController extends Controller
 {
+    private $validations = [
+        'slug'      => [
+            'required',
+            'string',
+            'max:100',
+        ],
+        'title'         => 'required|string|max:100',
+        'category_id'   => 'required|integer|exists:categories,id',
+        'tags'          => 'array',
+        'tags.*'        => 'integer|exists:tags,id',
+        'image'         => 'url|max:100',
+        'uploaded_img'  => 'nullable|image|max:1024',
+        'content'       => 'nullable|string',
+        'excerpt'       => 'nullable|string',
+    ];
     /**
      * Display a listing of the resource.
      *
@@ -35,9 +51,11 @@ class PostController extends Controller
     public function create()
     {
         $categories = Category::all('id', 'name');
+        $tags       = Tag::all();
 
         return view('admin.posts.create', [
             'categories'    => $categories,
+            'tags'          => $tags,
         ]);
     }
 
@@ -49,29 +67,24 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'title'         => 'required|string|max:100',
-            'slug'          => 'required|string|max:100|unique:posts',
-            'category_id'   => 'required|integer|exists:categories,id',
-            'image'         => 'url|max:100',
-            'uploaded_img'  => 'image|max:1024',
-            'content'       => 'string',
-            'excerpt'       => 'string',
-        ]);
+        $this->validations['slug'][] = 'unique:posts';
+        $request->validate($this->validations);
 
         $data = $request->all();
-
 
         $img_path = isset($data['uploaded_img']) ? Storage::put('uploads', $data['uploaded_img']) : null;
 
         $post = new Post;
-        $post->slug          = $data['slug'];
-        $post->title         = $data['title'];
-        $post->image         = $data['image'];
-        $post->uploaded_img  = $img_path;
-        $post->content       = $data['content'];
-        $post->excerpt       = $data['excerpt'];
+        $post->slug             = $data['slug'];
+        $post->title            = $data['title'];
+        $post->category_id      = $data['category_id'];
+        $post->image            = $data['image'];
+        $post->uploaded_img     = $img_path;
+        $post->content          = $data['content'];
+        $post->excerpt          = $data['excerpt'];
         $post->save();
+
+        $post->tags()->attach($data['tags']);
 
         return redirect()->route('admin.posts.show', ['post' => $post]);
     }
@@ -95,9 +108,14 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
-        {
-            return view('admin.posts.edit', compact('post'));
-        }
+        $categories = Category::all('id', 'name');
+        $tags       = Tag::all();
+
+        return view('admin.posts.edit', [
+            'post'          => $post,
+            'categories'    => $categories,
+            'tags'          => $tags,
+        ]);
     }
 
     /**
@@ -109,37 +127,29 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
-        {
+        $this->validations['slug'][] = Rule::unique('posts')->ignore($post);
+        $request->validate($this->validations);
 
-            $request->validate([
-                'slug'      => [
-                    'required',
-                    'string',
-                    'max:100',
-                    Rule::unique('posts')->ignore($post),
-                ],
-                'title'     => 'required|string|max:100',
-                'image'     => 'url|max:100',
-                'uploaded_img'  => 'image|max:1024',
-                'content'   => 'string',
-                'excerpt'   => 'string',
-            ]);
+        $data = $request->all();
 
-            $data = $request->all();
-
+        if (isset($data['uploaded_img'])) {
             $img_path = Storage::put('uploads', $data['uploaded_img']);
             Storage::delete($post->uploaded_img);
-
-            $post->slug     = $data['slug'];
-            $post->title    = $data['title'];
-            $post->image    = $data['image'];
-            $post->uploaded_img  = $img_path;
-            $post->content  = $data['content'];
-            $post->excerpt  = $data['excerpt'];
-            $post->update();
-
-            return redirect()->route('admin.posts.show', ['post' => $post]);
+        } else {
+            $img_path = $post->uploaded_img;
         }
+
+        $post->slug         = $data['slug'];
+        $post->title        = $data['title'];
+        $post->image        = $data['image'];
+        $post->uploaded_img  = $img_path;
+        $post->content  = $data['content'];
+        $post->excerpt  = $data['excerpt'];
+        $post->update();
+
+        $post->tags()->sync($data['tags']);
+
+        return redirect()->route('admin.posts.show', ['post' => $post]);
     }
 
     /**
@@ -151,6 +161,7 @@ class PostController extends Controller
     public function destroy(Post $post)
     {
         {
+            $post->tags()->sync([]);
             $post->delete();
 
             return redirect()->route('admin.posts.index')->with('success_delete', $post);
